+++ src/main/java/com/brinkcommerce/api/management/order/delivery/model/response/GiftCardProduct.java
@@ -3,7 +3,7 @@ package com.brinkcommerce.api.management.order.delivery.model.response;
 public record GiftCardProduct(
         String giftCardProductId,
         String currencyCode,
-        Integer priceAmount,
+        Long priceAmount,
         String status,
         String createTransactionId,
         String createFailedReason,
